The great refactor
